@@ -2,16 +2,18 @@
 import React from 'react';
 import _ from 'lodash';
 import PropTypes from 'prop-types';
-import { withStyles } from 'material-ui/styles';
-import List, { ListItem, ListItemText, ListItemSecondaryAction } from 'material-ui/List';
-import Checkbox from 'material-ui/Checkbox';
-import Avatar from 'material-ui/Avatar';
-import Paper from 'material-ui/Paper';
-import Dialog, { DialogTitle, DialogContent, DialogActions } from 'material-ui/Dialog';
-import Input from 'material-ui/Input';
-import Button from 'material-ui/Button';
-import Slide from 'material-ui/transitions/Slide';
+import { withStyles } from '@material-ui/core/styles';
+import { List, ListItem, ListItemText, ListItemSecondaryAction } from '@material-ui/core';
+import Checkbox from '@material-ui/core/Checkbox';
+import Avatar from '@material-ui/core/Avatar';
+import Paper from '@material-ui/core/Paper';
+import { Dialog,  DialogTitle, DialogContent, DialogActions } from '@material-ui/core';
+import Input from '@material-ui/core/Input';
+import Select from '@material-ui/core/Select';
+import Button from '@material-ui/core/Button';
+import Slide from '@material-ui/core/Slide';
 import { Test } from 'meteor/duckdodgerbrasl:lern-model';
+import { MenuItem } from '@material-ui/core';
 
 function Transition(props) {
   return <Slide direction="up" {...props} />;
@@ -91,7 +93,7 @@ class AdminTestScores extends React.Component {
 
     return (
       <div>
-        
+
         <List className={classes.root}>
 
           {
@@ -120,9 +122,9 @@ class AdminTestScores extends React.Component {
           aria-labelledby='Score a tag'
           open={open}
         >
-          
+
           <DialogTitle id='scores-title'>Set score proportion</DialogTitle>
-          
+
           <DialogContent>
 
             <Input
@@ -143,7 +145,7 @@ class AdminTestScores extends React.Component {
             <Button onClick={this.handleClose('cancel')} color="secondary">
               Cancel
             </Button>
-            
+
             <Button onClick={this.handleClose('save')} color="primary">
               Save
             </Button>
@@ -151,7 +153,7 @@ class AdminTestScores extends React.Component {
           </DialogActions>
 
         </Dialog>
-      
+
       </div>
     );
   }
