@@ -1,11 +1,11 @@
 // Libs
 import React from 'react';
 import _ from 'lodash';
-import List, { ListItem, ListItemText, ListItemSecondaryAction } from 'material-ui/List';
-import Avatar from 'material-ui/Avatar';
-import Paper from 'material-ui/Paper';
-import IconButton from 'material-ui/IconButton';
-import { Delete } from 'material-ui-icons';
+import { List, ListItem, ListItemText, ListItemSecondaryAction } from '@material-ui/core';
+import Avatar from '@material-ui/core/Avatar';
+import Paper from '@material-ui/core/Paper';
+import IconButton from '@material-ui/core/IconButton';
+import { Delete } from '@material-ui/icons';
 
 class AdminTestsList extends React.Component {
 
