@@ -1,8 +1,8 @@
 // Libs
 import React from 'react';
 import _ from 'lodash';
-import List, { ListItem, ListItemText } from 'material-ui/List';
-import Avatar from 'material-ui/Avatar';
+import { List, ListItem, ListItemText } from '@material-ui/core';
+import Avatar from '@material-ui/core/Avatar';
 
 class AdminUsersList extends React.Component {
 
