@@ -1,11 +1,11 @@
 import React from 'react';
-import TextField from 'material-ui/TextField';
-import Button from 'material-ui/Button';
+import TextField from '@material-ui/core/TextField';
+import Button from '@material-ui/core/Button';
 import Dialog, {
   DialogTitle,
   DialogActions,
   DialogContent,
-  DialogContentText } from 'material-ui/Dialog';
+  DialogContentText } from '@material-ui/core/Dialog';
 import { Layout } from 'meteor/duckdodgerbrasl:lern-layouts';
 import _ from 'lodash';
 
