@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { withStyles } from 'material-ui/styles';
-import { Typography, Button } from 'material-ui';
+import { withStyles } from '@material-ui/core/styles';
+import { Typography, Button } from '@material-ui/core';
 import { Sudoku } from 'meteor/duckdodgerbrasl:lern-games';
 import { Layout } from 'meteor/duckdodgerbrasl:lern-layouts';
 
