@@ -4,16 +4,16 @@ import _ from 'lodash';
 import log from 'loglevel';
 import { Content } from 'meteor/duckdodgerbrasl:lern-model';
 import PropTypes from 'prop-types';
-// Material Components
-import List, { ListItem, ListItemText } from 'material-ui/List';
-import { withStyles } from 'material-ui/styles';
-import Avatar from 'material-ui/Avatar';
-import Paper from 'material-ui/Paper';
-import Grid from 'material-ui/Grid';
-import MobileStepper from 'material-ui/MobileStepper';
-import Button from 'material-ui/Button';
-import KeyboardArrowLeft from 'material-ui-icons/KeyboardArrowLeft';
-import KeyboardArrowRight from 'material-ui-icons/KeyboardArrowRight';
+import { List, ListItem, ListItemText } from '@material-ui/core';
+import { withStyles } from '@material-ui/core/styles';
+import Avatar from '@material-ui/core/Avatar';
+import Paper from '@material-ui/core/Paper';
+import Grid from '@material-ui/core/Grid';
+import { Content } from 'meteor/duckdodgerbrasl:lern-model';
+import MobileStepper from '@material-ui/core/MobileStepper';
+import Button from '@material-ui/core/Button';
+import KeyboardArrowLeft from '@material-ui/icons//KeyboardArrowLeft';
+import KeyboardArrowRight from '@material-ui/icons//KeyboardArrowRight';
 
 // Views
 import StudentTestAttemptSudoku from './Sudoku.jsx';
