@@ -1,11 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { withStyles } from 'material-ui/styles';
-import AppBar from 'material-ui/AppBar';
-import Toolbar from 'material-ui/Toolbar';
-import Typography from 'material-ui/Typography';
-import IconButton from 'material-ui/IconButton';
-import Icon from 'material-ui/Icon';
+import { withStyles } from '@material-ui/core/styles';
+import AppBar from '@material-ui/core/AppBar';
+import Toolbar from '@material-ui/core/Toolbar';
+import Typography from '@material-ui/core/Typography';
+import IconButton from '@material-ui/core/IconButton';
+import Icon from '@material-ui/core/Icon';
 
 const styles = {
   flex: {
@@ -24,8 +24,8 @@ const styles = {
 };
 
 const getTitle = ({ title, crumbs }) =>
-  <div>
-    <span>
+  [
+    <span key='crumb'>
       {
         _.flatten(
           _.map(crumbs, (c, i) =>
@@ -44,10 +44,9 @@ const getTitle = ({ title, crumbs }) =>
           )
         )
       }
-
-    </span>
-    <span>{title}</span>
-  </div>;
+    </span>,
+    <span key='title'>{title}</span>,
+  ];
 
 /**
  * React Component for top bar
@@ -96,7 +95,7 @@ class Bar extends React.Component {
             <Icon className={classes[`${color}IconButton`]}>menu</Icon>
           </IconButton>
 
-          <Typography type='title' color='inherit' className={classes.flex}>
+          <Typography variant='title' color='inherit' className={classes.flex}>
             {getTitle({ title, crumbs })}
           </Typography>
 
