@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { withStyles } from 'material-ui/styles';
-import { Chip, Grid, TextField } from 'material-ui';
+import { withStyles } from '@material-ui/core/styles';
+import { Chip, Grid, TextField } from '@material-ui/core';
 import _ from 'lodash';
 
 import AutoComplete from './AutoComplete.jsx';
