@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { withStyles } from 'material-ui/styles';
-import { Button, Typography } from 'material-ui';
+import { withStyles } from '@material-ui/core/styles';
+import { Button, Typography } from '@material-ui/core';
 
 const styles = theme => ({
 
